refactor: UI 내에서 반복해서 쓰는 함수는 유틸 함수로 분리
