+++ src/view/ui/utils/createElement.js
@@ -1,0 +1,5 @@
+export default function createElement(type, text) {
+  const element = document.createElement(type);
+  element.innerText = text;
+  return element;
+}
