+++ src/view/ui/utils/toggleClassName.js
@@ -1,0 +1,6 @@
+export default function toggleClassName(element, className) {
+  if (element.classList.contains(className)) {
+    return element.classList.remove(className);
+  }
+  return element.classList.add(className);
+}
